Description:
This level is located in a area were random people are given a partner to team up with and thrown into the arena to face off the other duo. The map is not a huge so it should be easy for the players to understand and get a field to the arena. The arena is split into three levels, the floor the players are in is the biggest one compared to the other floors. The reason why there is that much space is so the players can travel in different path ways and battle in different areas in the map. The map has health packs, ammunition boxes, and some armor that are given for the players when the match starts, BUT there is a slightly better armor in the arena the players will have to race each to get, this is in the center of the level up in the third floor. There is also a explosive weapon located in the first floor that the players will also have to race each other to take. This is meant to be a fun a competitive map.